+++ src/Setting/Settings.php
@@ -61,6 +61,7 @@ final class Settings
     public const VAULTING_ENABLED_WALLET = self::SYSTEM_CONFIG_DOMAIN . 'vaultingEnabledWallet';
     public const VAULTING_ENABLED_ACDC = self::SYSTEM_CONFIG_DOMAIN . 'vaultingEnabledACDC';
     public const VAULTING_ENABLED_VENMO = self::SYSTEM_CONFIG_DOMAIN . 'vaultingEnabledVenmo';
+    public const IS_LOCAL_ENVIRONMENT = self::SYSTEM_CONFIG_DOMAIN . 'isLocalEnvironment';
 
     /**
      * @internal these may change at any time
@@ -98,6 +99,7 @@ final class Settings
         self::VAULTING_ENABLED_VENMO => false,
         self::CROSS_BORDER_MESSAGING_ENABLED => false,
         self::CROSS_BORDER_BUYER_COUNTRY => null,
+        self::IS_LOCAL_ENVIRONMENT => false,
     ];
 
     public const LIVE_CREDENTIAL_KEYS = [