+++ tests/Checkout/ExpressCheckout/SalesChannel/ExpressCreateOrderRouteTest.php
@@ -14,6 +14,8 @@ use Shopware\Core\Checkout\Cart\LineItem\LineItem;
 use Shopware\Core\Checkout\Cart\SalesChannel\CartService;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
+use Shopware\Core\System\SystemConfig\SystemConfigService;
+use Shopware\PayPalSDK\Struct\V2\Order;
 use Swag\PayPal\Checkout\Cart\Service\CartPriceService;
 use Swag\PayPal\Checkout\Exception\OrderZeroValueException;
 use Swag\PayPal\Checkout\ExpressCheckout\SalesChannel\ExpressCreateOrderRoute;
@@ -47,7 +49,7 @@ class ExpressCreateOrderRouteTest extends TestCase
     use GatewayTestBehaviour;
     use IntegrationTestBehaviour;
 
-    public function testCreatePayment(): void
+    public function testCreatePaymentWithZeroValueCart(): void
     {
         $salesChannelContext = $this->getSalesChannelContext();
 
@@ -62,6 +64,7 @@ class ExpressCreateOrderRouteTest extends TestCase
             $this->getContainer()->get(PayPalOrderBuilder::class),
             new OrderResource(self::orderGateway(), new ApiContextFactoryMock()),
             $this->getContainer()->get(CartPriceService::class),
+            $this->getContainer()->get(SystemConfigService::class),
             $this->createMock(RouterInterface::class),
             new NullLogger(),
         );
@@ -71,7 +74,7 @@ class ExpressCreateOrderRouteTest extends TestCase
         $route->createPayPalOrder(new Request(), $salesChannelContext);
     }
 
-    public function testCreatePaymentWithZeroValueCart(): void
+    public function testCreatePayment(): void
     {
         $salesChannelContext = $this->getSalesChannelContext();
 
@@ -79,13 +82,42 @@ class ExpressCreateOrderRouteTest extends TestCase
 
         static::assertSame(Response::HTTP_OK, $response->getStatusCode());
         static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        $request = $this->getClient()->getLast();
+        static::assertNotNull($request);
+
+        $order = (new Order())->assign($request->getRequestBody() ?? []);
+
+        $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+        static::assertNotNull($experienceContext);
+        static::assertNotNull($experienceContext->getOrderUpdateCallbackConfig());
     }
 
-    private function createRoute(): ExpressCreateOrderRoute
+    public function testCreateWithoutShippingCallback(): void
+    {
+        $salesChannelContext = $this->getSalesChannelContext();
+
+        $response = $this->createRoute(true)->createPayPalOrder(new Request(), $salesChannelContext);
+
+        static::assertSame(Response::HTTP_OK, $response->getStatusCode());
+        static::assertSame(CreateOrderCapture::ID, $response->getToken());
+
+        $request = $this->getClient()->getLast();
+        static::assertNotNull($request);
+
+        $order = (new Order())->assign($request->getRequestBody() ?? []);
+
+        $experienceContext = $order->getPaymentSource()?->getPaypal()?->getExperienceContext();
+        static::assertNotNull($experienceContext);
+        static::assertNull($experienceContext->getOrderUpdateCallbackConfig());
+    }
+
+    private function createRoute(bool $callbacksDisabled = false): ExpressCreateOrderRoute
     {
         $systemConfig = $this->createSystemConfigServiceMock([
             Settings::CLIENT_ID => 'testClientId',
             Settings::CLIENT_SECRET => 'testClientSecret',
+            Settings::IS_LOCAL_ENVIRONMENT => $callbacksDisabled,
         ]);
 
         $priceFormatter = new PriceFormatter();
@@ -108,6 +140,7 @@ class ExpressCreateOrderRouteTest extends TestCase
             $paypalOrderBuilder,
             new OrderResource(self::orderGateway(), new ApiContextFactoryMock()),
             $this->getContainer()->get(CartPriceService::class),
+            $systemConfig,
             $this->createMock(RouterInterface::class),
             new NullLogger(),
         );